requirements.txt
asgiref==3.2.10
Django==3.0.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-vote==2.2.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
